+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub(crate) mod auth;
 pub(crate) mod fairings;
 pub(crate) mod google;
 pub(crate) mod models;
