+++ src/models/jwt.rs
@@ -1,0 +1,4 @@
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Deserialize, Serialize)]
+pub struct JwtToken {}
